Fix first Firefox download

+++ demo/selenium-init.js
@@ -66,7 +66,9 @@ request(firefoxBaseURL + 'test_packages.json', function(error, response, body) {
 
   if (version > firefoxVersion) {
     fs.writeFileSync(firefoxVersionFile, version, 'utf8');
-    fs.unlinkSync(path.join(destDir, util.format(firefoxFileNameFmt, firefoxVersion)));
+    if (firefoxVersion !== -Infinity) {
+      fs.unlinkSync(path.join(destDir, util.format(firefoxFileNameFmt, firefoxVersion)));
+    }
   }
 
   var firefoxFileName = util.format(firefoxFileNameFmt, version);
